ai_service/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
python-dotenv
requests
langchain
faiss-cpu
sentence-transformers
numpy
